ovn: Process dhcp packet-ins and respond through packet-outs
The DHCP packets can be of two types
(1) DHCP Discover
(2) DHCP Request

For (1), the controller should respond with DHCP offer and for (2),
either DHCP Ack or DHCP Nack should be sent. In this patch, DHCP Nack
is never sent. In case of failures in validating the packet, the
controller does not respond at all.

For a DHCP packet, the IP address is read from the port binding table
with the source MAC as the search reference. The DHCP options for netmask
and router are expected with key values dhcp_opt_1 and
dhcp_opt_3 respectively in port binding table - options column.
In case of the absence of these options from
logical port entry, dhcp packet is flooded.

Signed-off-by: Babu Shanmugam <bschanmu@redhat.com>
Signed-off-by: Numan Siddique <nusiddiq@redhat.com>
Co-Authored-by: Numan Siddique <nusiddiq@redhat.com>

Conflicts:
	ovn/controller/pinctrl.c