ovn: Move setting of chassis column of port binding to ovn-northd
ovn-northd will set the chassis column of port binding table if the
CMS has defined the chassis id hosting the logial port's VIF in the
"Logical_Switch_Port.options:chassis" column.

ovn-controller will check the chassis column and add the necessary
OF flows for the VIF's present in the integration bridge.

One side effect of this is that the Logical_Switch_Port.up may not
be reflecting the true status.

Signed-off-by: Numan Siddique <nusiddiq@redhat.com>