ofproto-dpif: Fix the clone issue with continuation
When the clone action is composed and if the inner clone actions
modifies any of the metadata fields, the updated values are not
preserved when xlate_clone() returns. This causes the controller
to receive invalid metadata if the cloned packet has a 'controller'
action with the 'pause' flag set. When the controller resumes the
packet the resumed packet will not continue properly.

This patch addresses this issue.

One can reproduce the issue by running the test case added in this
patch without the fix.

Signed-off-by: Numan Siddique <nusiddiq@redhat.com>