ovn: Fix DHCP classless static route for non-classful masks.
When trying to determine how many bytes of ip address needs to be included
in classless static route option, we should take into consideration the
following.  To get the correct amount of bytes we need to take number of
network bits in the mask and divide it by 8.  But if the mask has a
remainder when divided, we need to not ignore this and add 1 byte to the to
the length of the option.

Signed-off-by: Rostyslav Fridman <rostyslav_fridman@epam.com>
Signed-off-by: Ben Pfaff <blp@ovn.org>
Signed-off-by: Lorenzo Bianconi <lorenzo.bianconi@redhat.com>