[RFC] Add a new OVS action check_pkt_larger
This patch adds a new action 'check_pkt_larger' which checks if the
packet is larger than the given size and stores the result in the
destination register.

Usage: check_pkt_larger:len->REGISTER
Eg. match=...,actions=check_pkt_larger:1442->NXM_NX_REG0[0],next;

This patch makes use of the new datapath action - 'check_pkt_len'
which is still WIP. At the start of ovs-vswitchd, datapath is probed
for this action. If the datapath action is present, then 'check_pkt_larger'
makes use of this datapath action.

Datapath action 'check_pkt_len' takes 3 nlattr's - pkt_len, userspace condition and
nested userspace attr - OVS_ACTION_ATTR_USERSPACE. Datapath when executing this
action sends the packet to the userspace if -

      * userspace condition is 1 (which means greater) and the actual
        length of the packet is greater than the specified pkt_len.

      * userspace condition is 2 (which means lesser or equal) and
        the actual length of the packet is lesser than or equal to
        the specified pkt_len.

If the packet is sent to the userspace, remaining actions after 'check_pkt_len' are
not executed.

Let's say we have these flows added to an OVS bridge br-int

table=0, priority=100 in_port=1,ip,actions=check_pkt_larger:100->NXM_NX_REG0[0],resubmit(,1)
table=1, priority=200,in_port=1,ip,reg0=0x1/0x1 actions=output:3
table=1, priority=100,in_port=1,ip,actions=output:4

Suppose if a packet is received from in_port=1 of length 150, the packet is sent
to userspace (MISS_UPCALL). The action 'check_pkt_larger' will be translated as
  - check_pkt_len( <= 100 ? userspace(pid=<PID>,slow_path(check_pkt_len))
And ovs-vswitchd will add a dp flow as
  - check_pkt_len( <= 100 ? userspace(pid=<PID>,slow_path(check_pkt_len)), output: 3

What it means is that, datapath will check the packet length and if the packet length
is less than 100, it will upcall the packet, else it will continue with the actions. In
this case, the packet will be output to port 3. Any subsequent packets with length
greater than 100 will be handled in datapath itself until the dp flow expires.

If in case, a packet with length lesser or equal to 100 is received, datapath will
upcall the packet. ovs-vswitchd will now replace the old flow with a new dp flow
 - check_pkt_len( > 100 ? userspace(pid=<PID>,slow_path(check_pkt_len)), output: 4

In case, datapath doesn't support 'check_pkt_len' action, the OVS action
'check_pkt_larger' sets SLOW_ACTION so that datapath flow is not added.

This OVS action is intended to be used by OVN to check the packet length
and generate an ICMP packet with type 3, code 4 and next hop mtu
in the logical router pipeline if the MTU of the physical interface
is lesser than the packet length. More information can be found here [1]

TODO:
 - Add test case.
 - Change the action format from check_pkt_larger:len to check_pkt_larger(len)

Concerns: I have concerns regarding the approach I have to taken to replace the
old dp flow with the new dp flow if there is a change in packet length. I am
not sure if it's a good idea. Other approach could be to not replace the dp flow.
ovs-vswitchd can add dp flow like
 - check_pkt_len( > pkt_len ? userspace(pid=<PID>,slow_path(check_pkt_len)), <other actions>
<other actions> will always gets executed if the packet len is <= pkt_len. Datapath will
upcall the packet only if the packet length is greater than pkt_len.

Request to suggest a better name for the action in case 'check_pkt_larger'
seems odd.

[1] - https://mail.openvswitch.org/pipermail/ovs-discuss/2018-July/047039.html

Reported-at:
https://mail.openvswitch.org/pipermail/ovs-discuss/2018-July/047039.html
Suggested-by: Ben Pfaff <blp@ovn.org>
Signed-off-by: Numan Siddique <nusiddiq@redhat.com>
CC: Ben Pfaff <blp@ovn.org>