ovn: Add generic HA chassis group
This patch adds the tables in both OVN Northbound and Southbound DBs to
support generic HA Chassis groups support in OVN. CMS can create
a group of HA chassis with the priorities assigned to each
chassis in the group. An HA chassis group can be associated to
a distributed logical router port or to 'external'* logical ports.

HA chassis group is similar to the present gateway chassis support in
OVN which is used by the distributed gateway router ports.
This patch tries to abstract this so that, the HA chassis support
can be leveraged by not just distributed gateway router ports.

[*] - An upcoming patch will add the 'external' logical ports
support.
Another patch will be added which will replace the gateway chassis
with the HA chassis group and will try to simplify the existing
gateway chassis code in ovn-controller.

Signed-off-by: Numan Siddique <nusiddiq@redhat.com>