ovn: Build HA chassis reference list for the HA chassis group
This patch updates the 'ref_chassis' column of HA Chassis groups
in the Southbound Db.

'ref_chassis' column of HA_Chassis_Group table is used to store
the list of chassis which references the HA chassis group. This
information will be used by ovn-controller in an upcoming patch
to establish BFD sessions with the required chassis.

Suppose if there is an HA chassis group - 'hagrp1' in the Southbound
DB and it has HA chasiss list - ha1, ha2 and ha3 and this HA chassis
group is used by a distributed logical router port, then ovn-northd
will update the 'ref_chassis' with the list of chassis which has claimed
all the logical switch ports which are connected to the logical router
to which the distributed logical router port belongs to.

Signed-off-by: Numan Siddique <nusiddiq@redhat.com>