ovn-northd: Delete the references to gateway_chasss in SB DB
Previous patch in the series add the support in ovn-controller
to use ha_chassis_group table in SB DB to support HA chassis
and establishing BFD tunnels instead of the gateway_chassis table.
There is no need for ovn-northd to create any gateway_chassis
rows in SB DB. This patch does that and deletes the code
which is not required anymore.

Signed-off-by: Numan Siddique <nusiddiq@redhat.com>