ovn-controller: Make use of ha_chassis_group table to bind the chassisredirect ports
This patch uses the newly added ha_chassis_group table in Southbound DB

 - to bind the chassisredirect ports.

 - to establish BFD sessions with the required chassis. The previous patch
   in this series sets the list of chassis which references a ha chassis group
   in the 'ref_chassis' column of 'ha_chassis_group' table (in ovn-northd).
   This patch uses that information to establish BFD sessions with only the
   required chassis. There is no need to traverse the local_datapath list
   to determine if a local chasis has to establish a BFD session with another
   chassis. For eg, if chassis - HV1, HV2 and HV3 are part of a chassis group
   G1 and G1 is referenced by compute chassis - C1 and C2, the chassis C1
   will establish BFD sessions with HV1, HV2 and HV3 since C1 references the
   group G1. The ha chassis HV1, HV2 and HV3 also establish BFD sessions
   amongst themselves and also with C1 and C2.

This patch also deletes the old code (which used gateway_chassis table)
to bind the chassisredirect port.

The rational behind the refactor is to make the ha chassis binding support
generic, so that logical ports of type 'external' (which will be
added in the upcoming patch) can also make use of it and to simplify
the gateway chassis support code in OVN. Functionally this new
approach is same as the older one.

Acked-by: Mark Michelson <mmichels@redhat.com>
Acked-by: Han Zhou <hzhou8@ebay.com>
Signed-off-by: Numan Siddique <nusiddiq@redhat.com>
Signed-off-by: Ben Pfaff <blp@ovn.org>
(cherry picked from commit 8256e80b828fec0cee62e6fa1ef5c90ebacd5181)